add SubMenuItem into SideBar
