+++ web/src/component/SideBar/SideBar.tsx
@@ -1,9 +1,62 @@
-import React from 'react';
+import React, { ReactNode, useState } from 'react';
 import { Link } from 'react-router-dom';
 import './SideBar.scss';
 import * as FiIcons from 'react-icons/fi';
+// import SubMenu from './SubMenu';
 import SideBarData from './SideBarData';
-import SubMenu from './SubMenu';
+
+interface CardItem {
+    item: ItemProps;
+}
+
+interface ItemProps {
+    path: string;
+    title: string;
+    icon: ReactNode;
+    subNav: {
+        path: string;
+        title: string;
+    }[];
+    iconOpened: ReactNode;
+    iconClosed: ReactNode;
+}
+
+const SubMenuItem: React.FC<CardItem> = function SubMenu({ item }) {
+    const [subnav, setSubnav] = useState(false);
+    const showSubnav = () => setSubnav(!subnav);
+
+    return (
+        <>
+            <Link
+                to={item.path}
+                className="sidebar-link"
+                onClick={item.subNav && showSubnav}
+            >
+                <div className="sidebar-item">
+                    {item.icon}
+                    <span className="menu-txt">{item.title}</span>
+                </div>
+                <div>
+                    {(() => {
+                        if (item.subNav.length > 0) {
+                            if (subnav) return item.iconOpened;
+                            return item.iconClosed;
+                        }
+                        return null;
+                    })()}
+                </div>
+            </Link>
+            {subnav &&
+                item.subNav.map((subItem) => {
+                    return (
+                        <Link to={subItem.path} className="dropdown-menu">
+                            <p>{subItem.title}</p>
+                        </Link>
+                    );
+                })}
+        </>
+    );
+};
 
 export default function SideBar() {
     return (
@@ -13,7 +66,7 @@ export default function SideBar() {
             </Link>
             <div className="sidebar-container">
                 {SideBarData.map((item) => {
-                    return <SubMenu {...item} />;
+                    return <SubMenuItem item={item} />;
                 })}
             </div>
